Implement joinGame
Check if the gameId supplied by a user effectively is the ID of
a game in the database, and check if the nickname the user
chose is not taken. If everything goes as expected,
create a new player instance, and attach the Id of the
new player to the response the client will receive.

If something goes wrong, an HTTP exception is raised.